RecLM-uni/requirements.txt
accelerate==0.25.0
bitsandbytes==0.40.2
datasets==2.15.0
FlagEmbedding==1.2.11
gradio==4.44.1
matplotlib==3.8.2
numpy==1.26.4
openai==1.3.7
pandas==2.2.1
peft==0.5.0
python_Levenshtein==0.23.0
PyYAML==6.0.1
Requests==2.32.3
sanic==23.12.1
seaborn==0.13.2
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2